:bug: Fix a code conflict caused by overriding the official Tray
